Fix network device creation logic.
Network device creation logic did not properly handle forwarded_ports;
instead it created connections to the management network for each one.
We depended on this in order to have at least one connection to the
management network. This commit explicitly handles the management
network and ignores forwarded ports.
