+++ lib/vagrant-libvirt/util.rb
@@ -4,7 +4,7 @@ module VagrantPlugins
       autoload :ErbTemplate, 'vagrant-libvirt/util/erb_template'
       autoload :Collection,  'vagrant-libvirt/util/collection'
       autoload :Timer,  'vagrant-libvirt/util/timer'
-      autoload :LibvirtUtil,  'vagrant-libvirt/util/libvirt_util'
+      autoload :NetworkUtil,  'vagrant-libvirt/util/network_util'
     end
   end
 end
